added_grades

+++ students/src/App/Http/Controllers/Students/Destroy.php
@@ -2,6 +2,8 @@
 
 namespace Sibtain\Students\App\Http\Controllers\Students;
 
+use LaravelEnso\Core\App\Models\User;
+use LaravelEnso\People\App\Models\Person;
 use Sibtain\Students\App\Student;
 use Illuminate\Routing\Controller;
 
@@ -9,7 +11,14 @@ class Destroy extends Controller
 {
     public function __invoke(Student $student)
     {
-        $student->delete();
+
+
+        if (User::where('email', $student->email)->delete()){
+            if (Person::where('email', $student->email)->delete())
+            {
+                $student->delete();
+            }
+        }
 
         return [
             'message' => __('The student was successfully deleted'),
