+++ teachers/src/App/Http/Controllers/Teachers/Destroy.php
@@ -2,15 +2,21 @@
 
 namespace Sibtain\Teachers\App\Http\Controllers\Teachers;
 
-use Sibtain\Teachers\App\Student;
+use LaravelEnso\Core\App\Models\User;
+use LaravelEnso\People\App\Models\Person;
+use Sibtain\Teachers\App\Teacher;
 use Illuminate\Routing\Controller;
 
 class Destroy extends Controller
 {
-    public function __invoke(Student $teacher)
+    public function __invoke(Teacher $teacher)
     {
-        $teacher->delete();
-
+        if (User::where('email', $teacher->email)->delete()){
+            if (Person::where('email', $teacher->email)->delete())
+            {
+                $teacher->delete();
+            }
+        }
         return [
             'message' => __('The teacher was successfully deleted'),
             'redirect' => 'teachers.index',
